+++ src/protocol/events/rtc-events.ts
@@ -1,45 +1,50 @@
 // tslint:disable:no-namespace
 // tslint:disable:max-classes-per-file
 // tslint:disable:ban-types
-// tslint:disable:member-ordering
-// tslint:disable:member-access
 import { DomainEvent } from './domain-event';
 
 export namespace rtcEvents {
 
   export abstract class RTCSignallingEvent implements DomainEvent {
+    public readonly callId: string;
+    public readonly sender: string;
+    public readonly tag: string;
+    public readonly __discriminator__ = 'domainEvent';
+
     constructor(callId: string, sender: string, tag: string) {
       this.callId = callId;
       this.sender = sender;
       this.tag = tag;
     }
 
-    readonly callId: string;
-    readonly sender: string;
-    readonly tag: string;
-    readonly __discriminator__ = 'domainEvent';
+    public static is = (e: DomainEvent): e is RTCSignallingEvent =>
+      typeof (e as RTCSignallingEvent).callId !== 'undefined'
   }
 
   export class DescriptionSent extends RTCSignallingEvent {
-    static readonly tag = 'rtc_description_sent';
+    public static readonly tag = 'rtc_description_sent';
+    public readonly sdp: RTCSessionDescriptionInit;
 
     constructor(callId: string, sender: string, sdp: RTCSessionDescriptionInit) {
       super(callId, sender, DescriptionSent.tag);
       this.sdp = sdp;
     }
 
-    readonly sdp: RTCSessionDescriptionInit;
+    public static is = (e: DomainEvent): e is DescriptionSent =>
+      e.tag === DescriptionSent.tag
   }
 
   export class CandidateSent extends RTCSignallingEvent {
-    static readonly tag = 'rtc_candidate_sent';
+    public static readonly tag = 'rtc_candidate_sent';
+    public readonly candidate: RTCIceCandidate;
 
     constructor(callId: string, sender: string, candidate: RTCIceCandidate) {
       super(callId, sender, CandidateSent.tag);
       this.candidate = candidate;
     }
 
-    readonly candidate: RTCIceCandidate;
+    public static is = (e: DomainEvent): e is CandidateSent =>
+      e.tag === CandidateSent.tag
   }
 
 }