+++ src/protocol/events/server-events.ts
@@ -1,14 +1,16 @@
 // tslint:disable:no-namespace
 // tslint:disable:max-classes-per-file
 // tslint:disable:ban-types
-// tslint:disable:member-ordering
-// tslint:disable:member-access
 import { DomainEvent } from './domain-event';
 
 export namespace serverEvents {
   export class Hello implements DomainEvent {
-    static readonly tag: string = 'hello';
-    readonly __discriminator__ = 'domainEvent';
+    public static readonly tag: string = 'hello';
+    public readonly __discriminator__ = 'domainEvent';
+    public readonly deviceId: string;
+    public readonly timestamp: number;
+    public readonly heartbeatTimeout: number;
+    public readonly tag = Hello.tag;
 
     constructor(deviceId: string, timestamp: number, heartbeatTimeout: number) {
       this.deviceId = deviceId;
@@ -16,21 +18,21 @@ export namespace serverEvents {
       this.heartbeatTimeout = heartbeatTimeout;
     }
 
-    readonly deviceId: string;
-    readonly timestamp: number;
-    readonly heartbeatTimeout: number;
-    readonly tag = Hello.tag;
+    public static is = (e: DomainEvent): e is Hello =>
+      e.tag === Hello.tag
   }
 
   export class OutputHeartbeat implements DomainEvent {
-    static readonly tag: string = 'output_heartbeat';
-    readonly __discriminator__ = 'domainEvent';
+    public static readonly tag: string = 'output_heartbeat';
+    public readonly __discriminator__ = 'domainEvent';
+    public readonly tag = OutputHeartbeat.tag;
+    public readonly timestamp: number;
 
     constructor(timestamp: number) {
       this.timestamp = timestamp;
     }
 
-    readonly tag = OutputHeartbeat.tag;
-    readonly timestamp: number;
+    public static is = (e: DomainEvent): e is OutputHeartbeat =>
+      e.tag === OutputHeartbeat.tag
   }
 }