+++ WHATS_NEW.md
@@ -1,4 +1,9 @@
 ### Improve
 
-* Simplify APIWithWebsocket class
-* Add optional Ref to errorEvents.Error to sync with Artichoke and avoid any casting
+* Simplify events - replace custom `EventHandler` with `rxjs`
+* Do not throw error if event was unhandled
+
+### Breaking changes
+* Callbacks registering was replaced by `Observables`
+  public callback methods were replaced from `onEventName` to `eventName$`
+